requirements.txt
streamlit>=1.25.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
mailjet_rest>=1.3.4
requests>=2.28.0
